feat(Entropy): Make generate_entropy more flexible with output
